pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "footy"
version = "12.0.0"
requires-python = ">=3.10"
description = "Football prediction system — v12 Analyst with 30+ expert council, LightGBM/CatBoost ensemble, and comprehensive data pipeline"
dependencies = [
  "httpx>=0.27,<1.0",
  "beautifulsoup4>=4.12,<5.0",
  "python-dotenv>=1.0,<2.0",
  "duckdb>=1.0,<2.0",
  "pandas>=2.2.3,<3.0",
  "numpy>=2.0,<3.0",
  "scipy>=1.11,<2.0",
  "scikit-learn>=1.6,<1.8",
  "typer>=0.12,<1.0",
  "rich>=13.7,<14.0",
  "gdeltdoc>=1.5,<2.0",
  "apscheduler>=3.10.4,<4.0",
  "fastapi>=0.109,<1.0",
  "uvicorn[standard]>=0.27,<1.0",
  "jinja2>=3.1,<4.0",
  "joblib>=1.3,<2.0",
  "lightgbm>=4.0",
  "catboost>=1.2",
  "optuna>=3.5",
  "soccerdata>=1.8",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0",
  "pytest-xdist>=3.5,<4.0",
  "httpx>=0.27,<1.0",
]

[project.scripts]
footy = "footy.cli:app"

[tool.setuptools]
package-dir = {"" = "src", "web" = "web"}
packages = [
  "footy",
  "footy.cli",
  "footy.models",
  "footy.models.experts",
  "footy.models.math",
  "footy.providers",
  "footy.llm",
  "web",
]

# ---------- pytest ----------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--tb=short",
  "-q",
]
markers = [
  "slow: marks tests as slow (deselect with '-m not slow')",
  "smoke: live-database smoke tests",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]

# ---------- coverage ----------
[tool.coverage.run]
source = ["footy", "web"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
  "pragma: no cover",
  "if __name__",
  "raise NotImplementedError",
]
